src/backend/requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.41.0
pydantic==2.12.5
pydantic-settings==2.13.1
python-dotenv==1.2.2
python-multipart==0.0.22
google-genai==1.66.0
pillow==12.1.1
